divide bookings into two
add choose to failing test
