+++ test/system/trips_workflow_test.rb
@@ -2,7 +2,7 @@
 
 require 'system_test_helper'
 
-class BookingWorkflowTest < ApplicationSystemTestCase
+class TripsWorkflowTest < ApplicationSystemTestCase
   setup do
     @user = FactoryBot.create(:user, :for_test, :with_couch, :subscribed)
     sign_in_as(@user)
@@ -119,6 +119,7 @@ class BookingWorkflowTest < ApplicationSystemTestCase
     assert_selector '.booking-form', count: 1
     assert_selector '.booking-form__button', count: 1
 
+    choose 'host'
     select_date('.booking_start_date', 3.days.from_now)
     select_date('.booking_end_date', 4.days.from_now)
     fill_in 'booking[message]', with: 'Hello, I would like to stay at your place.'
@@ -168,36 +169,5 @@ class BookingWorkflowTest < ApplicationSystemTestCase
 
     assert_selector '.upcoming-booking', count: 0
     assert_selector '.past-booking', count: 1
-  end
-
-  test 'should decline booking as host' do
-    visit request_booking_path(@request)
-    assert_no_selector 'button', text: 'Cancel Request'
-    click_on 'Decline Request'
-
-    assert_selector '.swal-modal'
-    assert_selector '.swal-title', text: 'Confirm'
-
-    click_on 'Yes, decline'
-
-    assert_selector 'div.flash', text: 'Request has been declined.'
-    assert_current_path requests_couch_bookings_path(@user.couch)
-
-    assert_selector '.upcoming-request', count: 0
-    assert_selector '.past-request', count: 1
-  end
-
-  test 'should accept booking as host' do
-    visit request_booking_path(@request)
-    assert_no_selector 'button', text: 'Update Request'
-
-    click_on 'Accept Request'
-
-    assert_selector '.swal-modal'
-    assert_selector '.swal-title', text: 'Request accepted!'
-    click_on 'Ok!'
-
-    assert_current_path request_booking_path(@request)
-    assert_selector '.booking__details-list > .status', text: 'Confirmed'
   end
 end
